+++ thebleep/shells/bash.py
@@ -24,7 +24,7 @@ class Bash(Generic):
         # the shell it ran in.
         return '''
             function {name} () {{
-                TB_EXIT=$?;
+                TB_EXIT=${{TB_EXIT:-$?}};
                 TB_SHELL_ALIASES=$(alias);
                 TB_HISTORY=$(fc -ln -10);
                 {fit_transport}